added nullable + various backend fixes

+++ ReadTrack.API/ReadTrack.API.Tests/Services/BaseServiceTests.cs
@@ -10,15 +10,46 @@ namespace ReadTrack.API.Tests.Services;
 
 public abstract class BaseServiceTests : IAsyncLifetime
 {
-    protected ReadTrackContext Context { get; private set; }
-    protected IMapper Mapper { get; private set; }
+    private ReadTrackContext? context;
+    private IMapper? mapper;
+    
+    protected ReadTrackContext Context
+    {
+        get
+        {
+            if (context == null)
+            {
+                throw new ArgumentNullException("Context not initialized");
+            }
+            return context;
+
+        }
+    }
+    
+    protected IMapper Mapper
+    {
+        get
+        {
+            if (mapper == null)
+            {
+                throw new ArgumentNullException("Mapper not initialized");
+            }
+            return mapper;
+
+        }
+    }
 
     protected virtual async Task SetupDbContextsAsync()
     {
         var builder = new DbContextOptionsBuilder<ReadTrackContext>();
         builder.UseInMemoryDatabase(Guid.NewGuid().ToString());
 
-        Context = (ReadTrackContext)Activator.CreateInstance(typeof(ReadTrackContext), builder.Options);
+        var contextInstance = Activator.CreateInstance(typeof(ReadTrackContext), builder.Options);
+        if (contextInstance is null)
+        {
+            throw new InvalidOperationException("Failed to create an instance of ReadTrackContext.");
+        }
+        context = (ReadTrackContext)contextInstance;
         await Context.Database.EnsureCreatedAsync();
         await Context.Database.EnsureDeletedAsync();
     }
@@ -26,7 +57,7 @@ public abstract class BaseServiceTests : IAsyncLifetime
     private void SetupMapper()
     {
         var configuration = new MapperConfiguration(cfg => cfg.AddProfile(new MappingProfile()));
-        Mapper = new Mapper(configuration);
+        mapper = new Mapper(configuration);
     }
 
     public async Task InitializeAsync()
